requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
qdrant-client==1.6.4
mysql-connector-python==8.2.0
pydantic~=2.7.2
sentence-transformers==2.2.2
python-magic==0.4.27
pillow==10.1.0
pytest==7.4.3
httpx==0.25.1
bcrypt==4.0.1
alembic==1.12.1
torch==2.5.0
jupyterlab==4.3.5
openai==1.64.0
pydantic-settings==2.8.0
pypdf==5.3.0
colpali-engine==0.3.8
orjson==3.10.15